+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_cosmos_client.py
@@ -276,17 +276,20 @@ class CosmosClient:  # pylint: disable=client-accepts-api-version-keyword
         session_token = kwargs.get('session_token')
         if session_token is not None:
             warnings.warn(
-                "The 'session_token' flag does not apply to this method and will be removed in the future.",
+                "The 'session_token' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         etag = kwargs.get('etag')
         if etag is not None:
             warnings.warn(
-                "The 'etag' flag does not apply to this method and will be removed in the future.",
+                "The 'etag' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         match_condition = kwargs.get('match_condition')
         if match_condition is not None:
             warnings.warn(
-                "The 'match_condition' flag does not apply to this method and will be removed in the future.",
+                "The 'match_condition' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         if initial_headers is not None:
             kwargs["initial_headers"] = initial_headers
@@ -327,17 +330,20 @@ class CosmosClient:  # pylint: disable=client-accepts-api-version-keyword
         session_token = kwargs.get('session_token')
         if session_token is not None:
             warnings.warn(
-                "The 'session_token' flag does not apply to this method and will be removed in the future.",
+                "The 'session_token' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         etag = kwargs.get('etag')
         if etag is not None:
             warnings.warn(
-                "The 'etag' flag does not apply to this method and will be removed in the future.",
+                "The 'etag' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         match_condition = kwargs.get('match_condition')
         if match_condition is not None:
             warnings.warn(
-                "The 'match_condition' flag does not apply to this method and will be removed in the future.",
+                "The 'match_condition' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         if initial_headers is not None:
             kwargs["initial_headers"] = initial_headers
@@ -391,7 +397,8 @@ class CosmosClient:  # pylint: disable=client-accepts-api-version-keyword
         session_token = kwargs.pop('session_token', None)
         if session_token is not None:
             warnings.warn(
-                "The 'session_token' flag does not apply to this method and will be removed in the future.",
+                "The 'session_token' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         if initial_headers is not None:
             kwargs["initial_headers"] = initial_headers
@@ -432,7 +439,8 @@ class CosmosClient:  # pylint: disable=client-accepts-api-version-keyword
         session_token = kwargs.get('session_token')
         if session_token is not None:
             warnings.warn(
-                "The 'session_token' flag does not apply to this method and will be removed in the future.",
+                "The 'session_token' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         if initial_headers is not None:
             kwargs["initial_headers"] = initial_headers
@@ -471,17 +479,20 @@ class CosmosClient:  # pylint: disable=client-accepts-api-version-keyword
         session_token = kwargs.get('session_token')
         if session_token is not None:
             warnings.warn(
-                "The 'session_token' flag does not apply to this method and will be removed in the future.",
+                "The 'session_token' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         etag = kwargs.get('etag')
         if etag is not None:
             warnings.warn(
-                "The 'etag' flag does not apply to this method and will be removed in the future.",
+                "The 'etag' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
         match_condition = kwargs.get('match_condition')
         if match_condition is not None:
             warnings.warn(
-                "The 'match_condition' flag does not apply to this method and will be removed in the future.",
+                "The 'match_condition' flag does not apply to this method and is always ignored even if passed."
+                " It will now be removed in the future.",
                 DeprecationWarning)
 
         if initial_headers is not None: